requirements.txt
mysql-connector==2.2.9
mysql-connector-python==8.2.0
pandas==2.1.1
streamlit==1.28.1
streamlit-authenticator==0.2.3

